requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
aisuite==0.1.11
alembic==1.13.2
alphashape==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.67.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
appnope==0.1.4
arxiv==2.2.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
browserbase==1.4.0
build==1.3.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==1.0.21
click==8.2.1
click-didyoumean==0.3.1
click-log==0.4.0
click-plugins==1.1.1.2
click-repl==0.3.0
cohere==5.18.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
ConfigArgParse==1.7.1
configparser==7.2.0
contourpy==1.3.3
courlan==1.3.2
Crawl4AI==0.3.745
crewai==0.186.1
crewai-tools==0.71.0
cryptography==45.0.7
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.2
ddgs==9.5.5
debugpy==1.8.16
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
diff-match-patch==20241021
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docling==2.52.0
docling-core==2.48.1
docling-ibm-models==3.9.1
docling-parse==4.4.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
easyocr==1.7.2
effdet==0.4.1
embedchain==0.1.128
emoji==2.14.1
et_xmlfile==2.0.0
exa-py==1.15.6
executing==2.2.1
fake-http-header==0.3.5
fake-useragent==2.2.0
Faker==37.6.0
fastapi==0.115.0
fastavro==1.12.0
fastembed==0.7.3
feedparser==6.0.12
filelock==3.18.0
filetype==1.2.0
flake8==7.3.0
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-vision==3.10.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.4
grep-ast==0.9.0
grpcio==1.74.0
grpcio-status==1.71.2
grpcio-tools==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.7
hpack==4.1.0
html2text==2024.2.26
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.34.3
humanfriendly==10.0
humanize==4.13.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==7.2.1
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.11.3
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jq==1.10.0
json5==0.12.0
json_repair==0.25.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jsonschema_pydantic==0.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jusText==3.0.2
kiwisolver==1.4.9
kombu==5.5.4
kubernetes==33.1.0
lance-namespace==0.0.6
lance-namespace-urllib3-client==0.0.14
lancedb==0.25.0
langchain==0.3.0
langchain-anthropic==0.3.20
langchain-chroma==0.2.6
langchain-cohere==0.3.5
langchain-community==0.3.29
langchain-core==0.3.76
langchain-docling==1.1.0
langchain-exa==0.3.1
langchain-experimental==0.3.4
langchain-openai==0.3.33
langchain-qdrant==0.2.1
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.3.45
lark==1.2.2
latex2mathml==3.78.1
lazy_loader==0.4
litellm==1.74.9
loguru==0.7.3
lxml==6.0.1
lxml_html_clean==0.4.2
Mako==1.3.10
markdown-it-py==3.0.0
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mccabe==0.7.0
mcp==1.1.3
mcp-use==1.2.6
mdurl==0.1.2
mem0ai==0.1.117
mindsdb_sdk==3.4.8
mindsdb_sql_parser==0.0.2
mixpanel==4.10.1
ml_dtypes==0.5.3
mmh3==5.2.0
monotonic==1.6
mpire==2.10.2
mpmath==1.3.0
mslex==1.3.0
multidict==6.6.3
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.1
nodeenv==1.9.1
notion-client==2.2.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.22.0
openai==1.107.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.3
ormsgpack==1.10.0
oslex==0.1.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
patchright==1.55.1
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pi_heif==1.1.0
pikepdf==9.11.0
pillow==10.4.0
platformdirs==4.4.0
playwright==1.55.0
pluggy==1.6.0
polyfactory==2.22.2
portalocker==2.7.0
postgrest==1.1.1
posthog==3.25.0
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
pyflakes==3.4.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
pylance==0.36.0
pylatexenc==2.10
PyNaCl==1.6.0
pyOpenSSL==25.1.0
pypandoc==1.15
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.405
pysbd==0.3.4
pytest==8.4.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==27.1.0
qdrant-client==1.15.1
rank-bm25==0.2.2
RapidFuzz==3.14.1
readability-lxml==0.8.1
realtime==2.7.0
redis==5.0.8
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.27.0
rsa==4.9.1
rtree==1.4.1
ruff==0.13.0
safetensors==0.6.2
schema==0.7.7
scikit-image==0.25.2
scipy==1.15.3
semchunk==2.2.2
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
slack_sdk==3.36.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
socksio==1.0.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.35
sse-starlette==2.0.0
sseclient-py==1.8.0
stack-data==0.6.3
StackAPI==0.3.1
stagehand==0.5.2
starlette==0.38.6
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
tf-playwright-stealth==1.2.0
tifffile==2025.9.9
tiktoken==0.10.0
timm==1.0.19
tld==0.13.1
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.56.1
tree-sitter==0.24.0
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.23.2
tree-sitter-language-pack==0.9.0
tree-sitter-yaml==0.7.1
trimesh==4.8.1
truststore==0.10.4
typer==0.16.1
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.18.14
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.17
uvicorn==0.30.6
uvloop==0.21.0
validators==0.20.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.3
xlsxwriter==3.2.6
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
